Refactor, additional getters/events, small bug fixes

+++ contracts/libraries/ByteArr.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.20;
 
 library ByteArr {
     function indexOf(bytes32[] storage self, bytes32 item) internal view returns (uint index, bool isThere) {
-        for (uint i; i < self.length; i++) {
+        for (uint256 i; i < self.length; i++) {
             if (self[i] == item) {
                 return (i, true);
             }
@@ -23,7 +23,7 @@ library ByteArr {
 
     function getFuncHash(bytes storage _data) internal view returns (bytes4) {
         bytes4 output;
-        for (uint i; i < 4; i++) {
+        for (uint256 i; i < 4; i++) {
             output |= bytes4(_data[i] & 0xFF) >> (i * 8);
         }
         return output;
